Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ long int l;
 long long int ll;
 float d;
 
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(i));
-printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(l));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(ll));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(d));
+printf("Size of a char: %d byte(s)\n", sizeof(c));
+printf("Size of a int: %d byte(s)\n", sizeof(i));
+printf("Size of a long int: %d byte(s)\n", sizeof(l));
+printf("Size of a long long int: %d byte(s)\n", sizeof(ll));
+printf("Size of a float: %d byte(s)\n", sizeof(d));
 return (0);
 }
